Improved formatPrettyPercentage functions
