+++ src/tools/PrizeTierController/utils/formatPrettyConfig.ts
@@ -1,13 +1,16 @@
 import { formatUnformattedBigNumberForDisplay } from '@pooltogether/utilities'
 import { PrizeTierConfigV2, PrettyConfig } from '@prizeTierController/interfaces'
-import { formatPrettyDprPercentage } from '@prizeTierController/utils/formatPrettyDprPercentage'
+import {
+  formatPrettyDprPercentage,
+  formatPrettyTierPercentage
+} from '@prizeTierController/utils/formatPrettyPercentage'
 
 export const formatPrettyConfig = (config: PrizeTierConfigV2, decimals: string): PrettyConfig => {
   const bitRangeSize = config.bitRangeSize.toLocaleString()
   const expiryDuration = config.expiryDuration.toLocaleString() + 's'
   const maxPicksPerUser = config.maxPicksPerUser.toLocaleString()
   const prize = formatUnformattedBigNumberForDisplay(config.prize, decimals)
-  const tiers = config.tiers.map((tier) => formatPrettyDprPercentage(tier))
+  const tiers = config.tiers.map((tier) => formatPrettyTierPercentage(tier))
   const endTimestampOffset = config.endTimestampOffset.toLocaleString() + 's'
   if (!!config.dpr) {
     const dpr = formatPrettyDprPercentage(config.dpr)
